+++ PowerShellForFOD/Public/Import-FODDynamicScan.ps1
@@ -74,8 +74,7 @@ function Import-FODDynamicScan {
             }
         }
         if (-not $ScanFile.Exists) {
-            Write-Error "Scan file '$ScanFile' does not exist"
-            throw
+            throw "Scan file '$ScanFile' does not exist"
         }
         $Params = @{}
         if ($Proxy) {